Hotfix for cw-mangle destructor crash
This isn't a full or proper fix. The issue stems from us trying to fill in a vtable with two different destructor types (as per Itanium), but mangling with CodeWarrior (only one destructor type) yields two unique functions with the same symbol. The ideal solution is support for CodeWarrior's universal destructors.